simulador_windows/requirements.txt
pymodbus==3.7.4
pyyaml==6.0.2
pyserial==3.5
